Fix TooManyRequestsException
https://github.com/KazaKago/swr-compose/issues/35
